lib: add a base class to enable new ones
the DSL was removed in 09b2f5788ff12e97473aba2ef7f70b2e6aa7b5f5
because it used OpenStruct which didn't scale on 2K+ files, but it
turns out people still want to explore the data.

Pave the way by restoring a Base class for the imminent Startup class
which will look just like the previous version without OpenStruct.
